requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-storages==1.14.2
boto3==1.33.6
psycopg2-binary==2.9.9
python-magic==0.4.27
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0